fix: fix linter error in Steps.tsx
